+++ lib/model/media/video/stereo_3d_modes.dart
@@ -1,0 +1,22 @@
+class Stereo3dModes {
+  static const names = {
+    'ab2l': 'above below half height left first',
+    'tb2l': 'above below half height left first',
+    'ab2r': 'above below half height right first',
+    'tb2r': 'above below half height right first',
+    'abl': 'above below left first',
+    'tbl': 'above below left first',
+    'abr': 'above below right first',
+    'tbr': 'above below right first',
+    'al': 'alternating frames left first',
+    'ar': 'alternating frames right first',
+    'sbs2l': 'side by side half width left first',
+    'sbs2r': 'side by side half width right first',
+    'sbsl': 'side by side left first',
+    'sbsr': 'side by side right first',
+    'irl': 'interleave rows left first',
+    'irr': 'interleave rows right first',
+    'icl': 'interleave columns left first',
+    'icr': 'interleave columns right first',
+  };
+}