+++ lib/services/common/services.dart
@@ -21,7 +21,6 @@ import 'package:aves_report_platform/aves_report_platform.dart';
 import 'package:aves_services/aves_services.dart';
 import 'package:aves_services_platform/aves_services_platform.dart';
 import 'package:aves_video/aves_video.dart';
-import 'package:aves_video_ffmpeg/aves_video_ffmpeg.dart';
 import 'package:aves_video_mpv/aves_video_mpv.dart';
 import 'package:get_it/get_it.dart';
 import 'package:path/path.dart' as p;
@@ -58,7 +57,7 @@ void initPlatformServices() {
   getIt.registerLazySingleton<AvesAvailability>(LiveAvesAvailability.new);
   getIt.registerLazySingleton<LocalMediaDb>(SqfliteLocalMediaDb.new);
   getIt.registerLazySingleton<AvesVideoControllerFactory>(MpvVideoControllerFactory.new);
-  getIt.registerLazySingleton<AvesVideoMetadataFetcher>(FfmpegVideoMetadataFetcher.new);
+  getIt.registerLazySingleton<AvesVideoMetadataFetcher>(MpvVideoMetadataFetcher.new);
 
   getIt.registerLazySingleton<AppService>(PlatformAppService.new);
   getIt.registerLazySingleton<AppProfileService>(PlatformAppProfileService.new);